Start refactoring into driver tasks
